Extract console content ownership
