+++ clients/apple/AlanNative/Views/Console/ContentView.swift
@@ -1,616 +1,4 @@
 import SwiftUI
-#if os(iOS)
-import UIKit
-#elseif os(macOS)
-import AppKit
-#endif
-
-@MainActor
-final class AlanConsoleViewModel: ObservableObject {
-    enum ComposeMode: String, CaseIterable, Identifiable {
-        case turn = "New Turn"
-        case steer = "Steer"
-
-        var id: String { rawValue }
-    }
-
-    enum SessionConnectionState {
-        case idle
-        case connecting
-        case online
-        case degraded
-
-        var label: String {
-            switch self {
-            case .idle:
-                return "Idle"
-            case .connecting:
-                return "Syncing"
-            case .online:
-                return "Online"
-            case .degraded:
-                return "Reconnecting"
-            }
-        }
-
-        var tint: Color {
-            switch self {
-            case .idle:
-                return .gray
-            case .connecting:
-                return .yellow
-            case .online:
-                return .green
-            case .degraded:
-                return .orange
-            }
-        }
-    }
-
-    @Published var baseURLString = "http://127.0.0.1:8090"
-    @Published var healthStatus = "Not checked"
-    @Published var lastError: String?
-
-    @Published var governanceProfile: GovernanceProfile = .conservative
-    @Published var streamingMode: SessionStreamingMode = .auto
-    @Published var partialStreamRecoveryMode: PartialStreamRecoveryMode = .continueOnce
-
-    @Published var sessions: [SessionListItem] = []
-    @Published var selectedSessionID: String?
-
-    @Published var composeMode: ComposeMode = .turn
-    @Published var draftMessage = ""
-
-    @Published var messages: [ChatMessage] = []
-    @Published var timeline: [TimelineEntry] = []
-
-    @Published var pendingYield: PendingYieldState?
-    @Published var confirmationModifications = ""
-    @Published var structuredAnswerDrafts: [String: String] = [:]
-    @Published var customResumePayload = "{\n  \"choice\": \"approve\"\n}"
-
-    @Published var rollbackTurnsText = "1"
-    @Published var connectionState: SessionConnectionState = .idle
-
-    @Published var isCheckingHealth = false
-    @Published var isRefreshingSessions = false
-    @Published var isCreatingSession = false
-    @Published var isSending = false
-    @Published var isRunningAction = false
-
-    private var lastEventID: String?
-    private var eventPumpTask: Task<Void, Never>?
-    private var eventReducer = ConsoleEventReducer()
-    private var lastPumpErrorMessage: String?
-
-    deinit {
-        eventPumpTask?.cancel()
-    }
-
-    func bootstrap() async {
-        await refreshHealth()
-        await refreshSessions(keepSelection: false)
-    }
-
-    func shutdown() {
-        eventPumpTask?.cancel()
-        eventPumpTask = nil
-    }
-
-    var canSendMessage: Bool {
-        !isSending && !draftMessage.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    func refreshHealth() async {
-        isCheckingHealth = true
-        defer { isCheckingHealth = false }
-
-        do {
-            let response = try await client().checkHealth()
-            healthStatus = response.status
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Health Check Failed")
-        }
-    }
-
-    func refreshSessions(keepSelection: Bool = true) async {
-        isRefreshingSessions = true
-        defer { isRefreshingSessions = false }
-
-        do {
-            let sessionList = try await client().listSessions()
-            sessions = sessionList
-            if keepSelection {
-                if let selectedSessionID,
-                   sessionList.contains(where: { $0.sessionID == selectedSessionID })
-                {
-                    return
-                }
-                if selectedSessionID != nil {
-                    selectedSessionID = nil
-                    messages = []
-                    pendingYield = nil
-                    eventReducer.reset()
-                    lastEventID = nil
-                    eventPumpTask?.cancel()
-                    eventPumpTask = nil
-                    connectionState = .idle
-                }
-            } else {
-                selectedSessionID = nil
-            }
-        } catch {
-            reportError(error, timelineTitle: "Session List Failed")
-        }
-    }
-
-    func createSession() async {
-        isCreatingSession = true
-        defer { isCreatingSession = false }
-
-        do {
-            let response = try await client().createSession(
-                governanceProfile: governanceProfile,
-                streamingMode: streamingMode,
-                partialStreamRecoveryMode: partialStreamRecoveryMode
-            )
-            appendTimeline(
-                "Session Created",
-                detail: shortID(response.sessionID),
-                level: .action
-            )
-            await refreshSessions(keepSelection: true)
-            try await activateSession(response.sessionID)
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Create Session Failed")
-        }
-    }
-
-    func openSession(_ sessionID: String) async {
-        do {
-            try await activateSession(sessionID)
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Open Session Failed")
-        }
-    }
-
-    func forkCurrentSession() async {
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-        isRunningAction = true
-        defer { isRunningAction = false }
-
-        do {
-            let response = try await client().forkSession(sessionID: sessionID)
-            appendTimeline(
-                "Session Forked",
-                detail: "\(shortID(response.forkedFromSessionID)) -> \(shortID(response.sessionID))",
-                level: .action
-            )
-            await refreshSessions(keepSelection: true)
-            try await activateSession(response.sessionID)
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Fork Failed")
-        }
-    }
-
-    func deleteCurrentSession() async {
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-        isRunningAction = true
-        defer { isRunningAction = false }
-
-        do {
-            try await client().deleteSession(sessionID: sessionID)
-            appendTimeline("Session Deleted", detail: shortID(sessionID), level: .action)
-            selectedSessionID = nil
-            messages = []
-            pendingYield = nil
-            eventReducer.reset()
-            lastEventID = nil
-            eventPumpTask?.cancel()
-            eventPumpTask = nil
-            connectionState = .idle
-            await refreshSessions(keepSelection: false)
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Delete Failed")
-        }
-    }
-
-    func sendDraftMessage() async {
-        let text = draftMessage.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !text.isEmpty else {
-            return
-        }
-
-        isSending = true
-        defer { isSending = false }
-
-        do {
-            let sessionID = try await ensureActiveSession()
-
-            switch composeMode {
-            case .turn:
-                appendUserMessage(text)
-                _ = try await client().sendTurn(sessionID: sessionID, text: text)
-                appendTimeline("Turn Submitted", detail: shortID(sessionID), level: .action)
-            case .steer:
-                _ = try await client().sendInput(sessionID: sessionID, text: text)
-                appendTimeline("Steer Input Submitted", detail: text, level: .action)
-            }
-
-            draftMessage = ""
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "Send Failed")
-        }
-    }
-
-    func interruptCurrentSession() async {
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-        await runSessionAction(title: "Interrupt", detail: shortID(sessionID)) {
-            _ = try await self.client().interrupt(sessionID: sessionID)
-        }
-    }
-
-    func compactCurrentSession() async {
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-        await runSessionAction(title: "Compaction Requested", detail: shortID(sessionID)) {
-            _ = try await self.client().compact(sessionID: sessionID)
-        }
-    }
-
-    func rollbackCurrentSession() async {
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-        let turns = Int(rollbackTurnsText) ?? 0
-        guard turns > 0 else {
-            lastError = "Rollback turns must be a positive integer"
-            return
-        }
-
-        await runSessionAction(title: "Rollback Requested", detail: "\(turns) turn(s)") {
-            _ = try await self.client().rollback(sessionID: sessionID, turns: turns)
-        }
-    }
-
-    func approvePendingConfirmation() async {
-        await submitConfirmation(choice: "approve", modifications: nil)
-    }
-
-    func rejectPendingConfirmation() async {
-        await submitConfirmation(choice: "reject", modifications: nil)
-    }
-
-    func modifyPendingConfirmation() async {
-        let modifications = confirmationModifications
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !modifications.isEmpty else {
-            lastError = "Modification note cannot be empty"
-            return
-        }
-        await submitConfirmation(choice: "modify", modifications: modifications)
-    }
-
-    func submitStructuredInputAnswers() async {
-        guard let pendingYield, pendingYield.kind == "structured_input" else {
-            return
-        }
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-
-        var answers: [JSONValue] = []
-        for question in pendingYield.questions {
-            let rawValue = structuredAnswerDrafts[question.id] ?? ""
-            let value = rawValue.trimmingCharacters(in: .whitespacesAndNewlines)
-            if question.required && value.isEmpty {
-                lastError = "Question '\(question.label)' is required"
-                return
-            }
-            if !value.isEmpty {
-                answers.append(
-                    .object([
-                        "question_id": .string(question.id),
-                        "value": .string(value),
-                    ])
-                )
-            }
-        }
-
-        let payload = JSONValue.object([
-            "answers": .array(answers),
-        ])
-
-        await runSessionAction(title: "Structured Input Submitted", detail: shortID(pendingYield.requestID)) {
-            _ = try await self.client().resume(
-                sessionID: sessionID,
-                requestID: pendingYield.requestID,
-                payload: payload
-            )
-            self.pendingYield = nil
-            self.structuredAnswerDrafts = [:]
-        }
-    }
-
-    func submitCustomResumePayload() async {
-        guard let pendingYield else {
-            return
-        }
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-
-        let payloadText = customResumePayload.trimmingCharacters(in: .whitespacesAndNewlines)
-        let payload: JSONValue
-        if payloadText.isEmpty {
-            payload = .object([:])
-        } else {
-            payload = parseResumePayload(from: payloadText)
-        }
-
-        await runSessionAction(title: "Yield Resumed", detail: pendingYield.kind) {
-            _ = try await self.client().resume(
-                sessionID: sessionID,
-                requestID: pendingYield.requestID,
-                payload: payload
-            )
-            self.pendingYield = nil
-        }
-    }
-
-    func clearTimeline() {
-        timeline = []
-    }
-
-    private func activateSession(_ sessionID: String) async throws {
-        selectedSessionID = sessionID
-        connectionState = .connecting
-        pendingYield = nil
-        confirmationModifications = ""
-        structuredAnswerDrafts = [:]
-        eventReducer.reset()
-        lastEventID = nil
-
-        let session = try await client().readSession(sessionID: sessionID)
-        messages = session.messages.map { history in
-            let role = chatRole(from: history.role)
-            let text: String
-            if let toolName = history.toolName, !toolName.isEmpty {
-                text = "[\(toolName)] \(history.content)"
-            } else {
-                text = history.content
-            }
-            return ChatMessage(role: role, text: text, turnID: nil, isStreaming: false)
-        }
-        timeline = []
-        appendTimeline("Session Loaded", detail: shortID(sessionID), level: .info)
-
-        _ = try? await client().resumeSession(sessionID: sessionID)
-
-        startEventPump(sessionID: sessionID)
-    }
-
-    private func ensureActiveSession() async throws -> String {
-        if let selectedSessionID {
-            return selectedSessionID
-        }
-
-        let response = try await client().createSession(
-            governanceProfile: governanceProfile,
-            streamingMode: streamingMode,
-            partialStreamRecoveryMode: partialStreamRecoveryMode
-        )
-        await refreshSessions(keepSelection: true)
-        try await activateSession(response.sessionID)
-        return response.sessionID
-    }
-
-    private func startEventPump(sessionID: String) {
-        eventPumpTask?.cancel()
-        eventPumpTask = Task { [weak self] in
-            await self?.runEventPump(sessionID: sessionID)
-        }
-    }
-
-    private func runEventPump(sessionID: String) async {
-        connectionState = .connecting
-        let eventReader: ConsoleEventReader
-        do {
-            eventReader = try ConsoleEventReader(client: client())
-        } catch {
-            reportEventPumpError(error)
-            return
-        }
-
-        while !Task.isCancelled {
-            guard selectedSessionID == sessionID else {
-                return
-            }
-
-            do {
-                let page = try await eventReader.readNextPage(
-                    sessionID: sessionID,
-                    afterEventID: lastEventID
-                )
-
-                connectionState = .online
-                lastPumpErrorMessage = nil
-
-                if page.gap {
-                    appendTimeline(
-                        "Event Gap Detected",
-                        detail: "Re-sync from available buffer",
-                        level: .warning
-                    )
-                }
-
-                if let oldest = page.oldestEventID, page.gap {
-                    lastEventID = oldest
-                }
-
-                if page.events.isEmpty {
-                    try await Task.sleep(nanoseconds: 250_000_000)
-                    continue
-                }
-
-                for event in page.events {
-                    if Task.isCancelled || selectedSessionID != sessionID {
-                        return
-                    }
-                    lastEventID = event.eventID
-                    reduce(event: event)
-                }
-            } catch is CancellationError {
-                return
-            } catch {
-                reportEventPumpError(error)
-                try? await Task.sleep(nanoseconds: 1_000_000_000)
-            }
-        }
-    }
-
-    private func reduce(event: SessionEventEnvelope) {
-        var projection = ConsoleEventProjection(
-            messages: messages,
-            timeline: timeline,
-            pendingYield: pendingYield,
-            confirmationModifications: confirmationModifications,
-            structuredAnswerDrafts: structuredAnswerDrafts,
-            customResumePayload: customResumePayload
-        )
-        eventReducer.reduce(event: event, into: &projection)
-        messages = projection.messages
-        timeline = projection.timeline
-        pendingYield = projection.pendingYield
-        confirmationModifications = projection.confirmationModifications
-        structuredAnswerDrafts = projection.structuredAnswerDrafts
-        customResumePayload = projection.customResumePayload
-    }
-
-    private func appendUserMessage(_ text: String) {
-        messages.append(ChatMessage(role: .user, text: text))
-    }
-
-    private func appendTimeline(_ title: String, detail: String?, level: TimelineEntry.Level) {
-        timeline.append(TimelineEntry(title: title, detail: detail, level: level))
-        if timeline.count > 800 {
-            timeline.removeFirst(timeline.count - 800)
-        }
-    }
-
-    private func submitConfirmation(choice: String, modifications: String?) async {
-        guard let pendingYield else {
-            return
-        }
-        guard pendingYield.kind == "confirmation" else {
-            return
-        }
-        guard let sessionID = selectedSessionID else {
-            return
-        }
-
-        var payload: [String: JSONValue] = [
-            "choice": .string(choice),
-        ]
-
-        if let modifications, !modifications.isEmpty {
-            payload["modifications"] = .string(modifications)
-        }
-
-        await runSessionAction(title: "Confirmation Submitted", detail: choice) {
-            _ = try await self.client().resume(
-                sessionID: sessionID,
-                requestID: pendingYield.requestID,
-                payload: .object(payload)
-            )
-            self.pendingYield = nil
-            self.confirmationModifications = ""
-        }
-    }
-
-    private func runSessionAction(
-        title: String,
-        detail: String,
-        operation: @escaping () async throws -> Void
-    ) async {
-        isRunningAction = true
-        defer { isRunningAction = false }
-
-        do {
-            try await operation()
-            appendTimeline(title, detail: detail, level: .action)
-            lastError = nil
-        } catch {
-            reportError(error, timelineTitle: "\(title) Failed")
-        }
-    }
-
-    private func parseResumePayload(from text: String) -> JSONValue {
-        if let data = text.data(using: .utf8),
-           let object = try? JSONSerialization.jsonObject(with: data)
-        {
-            return JSONValue.from(any: object)
-        }
-        return .string(text)
-    }
-
-    private func shortID(_ value: String?) -> String {
-        guard let value, !value.isEmpty else {
-            return "-"
-        }
-        if value.count <= 12 {
-            return value
-        }
-        let prefix = value.prefix(8)
-        return "\(prefix)…"
-    }
-
-    private func chatRole(from rawRole: String) -> ChatMessage.Role {
-        switch rawRole.lowercased() {
-        case "user":
-            return .user
-        case "assistant":
-            return .assistant
-        case "error":
-            return .error
-        default:
-            return .system
-        }
-    }
-
-    private func reportError(_ error: Error, timelineTitle: String) {
-        let message = error.localizedDescription
-        lastError = message
-        appendTimeline(timelineTitle, detail: message, level: .error)
-    }
-
-    private func reportEventPumpError(_ error: Error) {
-        connectionState = .degraded
-        let message = error.localizedDescription
-        if message != lastPumpErrorMessage {
-            lastPumpErrorMessage = message
-            appendTimeline("Event Stream Error", detail: message, level: .warning)
-            lastError = message
-        }
-    }
-
-    private func client() throws -> AlanAPIClient {
-        try AlanAPIClient(baseURLString: baseURLString)
-    }
-}
 
 struct ContentView: View {
     @StateObject private var viewModel = AlanConsoleViewModel()
@@ -1415,294 +803,6 @@ struct ContentView: View {
             set: { newValue in
                 viewModel.structuredAnswerDrafts[questionID] = newValue
             }
-        )
-    }
-}
-
-private extension Color {
-    static func adaptive(
-        light: (Double, Double, Double),
-        dark: (Double, Double, Double),
-        alpha: Double = 1
-    ) -> Color {
-        #if os(iOS)
-        return Color(
-            UIColor { traitCollection in
-                let rgb = traitCollection.userInterfaceStyle == .dark ? dark : light
-                return UIColor(red: rgb.0, green: rgb.1, blue: rgb.2, alpha: alpha)
-            }
-        )
-        #elseif os(macOS)
-        return Color(
-            NSColor(name: nil) { appearance in
-                let isDark = appearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
-                let rgb = isDark ? dark : light
-                return NSColor(red: rgb.0, green: rgb.1, blue: rgb.2, alpha: alpha)
-            }
-        )
-        #else
-        return Color(red: dark.0, green: dark.1, blue: dark.2, opacity: alpha)
-        #endif
-    }
-}
-
-private enum ConsoleTheme {
-    static let windowBackground = LinearGradient(
-        colors: [
-            Color.adaptive(light: (0.96, 0.97, 0.99), dark: (0.06, 0.07, 0.08)),
-            Color.adaptive(light: (0.92, 0.94, 0.97), dark: (0.09, 0.10, 0.12)),
-            Color.adaptive(light: (0.95, 0.96, 0.98), dark: (0.07, 0.08, 0.10)),
-        ],
-        startPoint: .topLeading,
-        endPoint: .bottomTrailing
-    )
-
-    static let sidebarFill = Color.adaptive(light: (0.95, 0.96, 0.98), dark: (0.13, 0.14, 0.16))
-    static let sidebarBorder = Color.adaptive(light: (0.82, 0.85, 0.90), dark: (0.29, 0.31, 0.36))
-
-    static let panelFill = Color.adaptive(light: (1.00, 1.00, 1.00), dark: (0.10, 0.11, 0.13))
-    static let inlinePanel = Color.adaptive(light: (0.95, 0.96, 0.98), dark: (0.14, 0.15, 0.18))
-    static let panelBorder = Color.adaptive(light: (0.84, 0.87, 0.92), dark: (0.28, 0.30, 0.34))
-
-    static let textPrimary = Color.adaptive(light: (0.08, 0.10, 0.14), dark: (0.94, 0.95, 0.97))
-    static let textSecondary = Color.adaptive(light: (0.22, 0.26, 0.33), dark: (0.74, 0.77, 0.83))
-    static let textMuted = Color.adaptive(light: (0.40, 0.45, 0.53), dark: (0.56, 0.60, 0.67))
-
-    static let accent = Color.adaptive(light: (0.20, 0.44, 0.85), dark: (0.28, 0.58, 0.96))
-    static let success = Color.adaptive(light: (0.17, 0.62, 0.33), dark: (0.39, 0.81, 0.48))
-    static let warning = Color.adaptive(light: (0.85, 0.56, 0.12), dark: (0.98, 0.75, 0.29))
-    static let error = Color.adaptive(light: (0.77, 0.22, 0.24), dark: (0.94, 0.38, 0.40))
-
-    static let selectionFill = Color.adaptive(light: (0.74, 0.84, 0.99), dark: (0.24, 0.48, 0.85))
-        .opacity(0.30)
-    static let selectionBorder = Color.adaptive(light: (0.33, 0.54, 0.90), dark: (0.34, 0.63, 0.99))
-        .opacity(0.58)
-
-    static let badgeFill = Color.adaptive(light: (0.74, 0.84, 0.99), dark: (0.24, 0.48, 0.85))
-        .opacity(0.30)
-    static let badgeBorder = Color.adaptive(light: (0.33, 0.54, 0.90), dark: (0.34, 0.63, 0.99))
-        .opacity(0.58)
-
-    static let userBubble = Color.adaptive(light: (0.86, 0.92, 1.00), dark: (0.16, 0.23, 0.34))
-    static let assistantBubble = Color.adaptive(light: (0.95, 0.96, 0.98), dark: (0.14, 0.15, 0.18))
-    static let systemBubble = Color.adaptive(light: (0.93, 0.94, 0.96), dark: (0.17, 0.17, 0.19))
-    static let errorBubble = Color.adaptive(light: (1.00, 0.91, 0.92), dark: (0.26, 0.13, 0.14))
-}
-
-private struct SidebarActionButtonStyle: ButtonStyle {
-    var prominent: Bool
-
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .font(.system(size: 13, weight: .semibold, design: .rounded))
-            .foregroundStyle(prominent ? Color.white : ConsoleTheme.textPrimary)
-            .padding(.horizontal, 12)
-            .padding(.vertical, 9)
-            .background(
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .fill(prominent ? ConsoleTheme.accent : ConsoleTheme.inlinePanel)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .strokeBorder(prominent ? Color.clear : ConsoleTheme.panelBorder, lineWidth: 1)
-            )
-            .opacity(configuration.isPressed ? 0.82 : 1)
-            .animation(.easeOut(duration: 0.15), value: configuration.isPressed)
-    }
-}
-
-private struct InlineActionButtonStyle: ButtonStyle {
-    var prominent: Bool = false
-
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .font(.system(size: 12, weight: .semibold, design: .rounded))
-            .foregroundStyle(prominent ? Color.white : ConsoleTheme.textPrimary)
-            .padding(.horizontal, 10)
-            .padding(.vertical, 7)
-            .background(
-                RoundedRectangle(cornerRadius: 9, style: .continuous)
-                    .fill(prominent ? ConsoleTheme.accent : ConsoleTheme.inlinePanel)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 9, style: .continuous)
-                    .strokeBorder(prominent ? Color.clear : ConsoleTheme.panelBorder, lineWidth: 1)
-            )
-            .opacity(configuration.isPressed ? 0.84 : 1)
-            .animation(.easeOut(duration: 0.15), value: configuration.isPressed)
-    }
-}
-
-private struct CompactDarkFieldStyle: TextFieldStyle {
-    func _body(configuration: TextField<Self._Label>) -> some View {
-        configuration
-            .font(.system(size: 12, weight: .regular, design: .rounded))
-            .foregroundStyle(ConsoleTheme.textPrimary)
-            .padding(.horizontal, 10)
-            .padding(.vertical, 7)
-            .background(ConsoleTheme.inlinePanel, in: RoundedRectangle(cornerRadius: 9, style: .continuous))
-            .overlay(
-                RoundedRectangle(cornerRadius: 9, style: .continuous)
-                    .strokeBorder(ConsoleTheme.panelBorder, lineWidth: 1)
-            )
-    }
-}
-
-private struct MessageBubble: View {
-    let message: ChatMessage
-
-    private var roleTitle: String {
-        switch message.role {
-        case .user:
-            return "You"
-        case .assistant:
-            return "Alan"
-        case .system:
-            return "System"
-        case .error:
-            return "Error"
-        }
-    }
-
-    private var indicatorColor: Color {
-        switch message.role {
-        case .user:
-            return ConsoleTheme.accent
-        case .assistant:
-            return ConsoleTheme.success
-        case .system:
-            return ConsoleTheme.textMuted
-        case .error:
-            return ConsoleTheme.error
-        }
-    }
-
-    private var bubbleFill: Color {
-        switch message.role {
-        case .user:
-            return ConsoleTheme.userBubble
-        case .assistant:
-            return ConsoleTheme.assistantBubble
-        case .system:
-            return ConsoleTheme.systemBubble
-        case .error:
-            return ConsoleTheme.errorBubble
-        }
-    }
-
-    private var bodyText: String {
-        if message.text.isEmpty && message.isStreaming {
-            return "…"
-        }
-        return message.text
-    }
-
-    var body: some View {
-        HStack {
-            if message.role == .user {
-                Spacer(minLength: 72)
-            }
-
-            VStack(alignment: .leading, spacing: 7) {
-                HStack(spacing: 6) {
-                    Circle()
-                        .fill(indicatorColor)
-                        .frame(width: 7, height: 7)
-
-                    Text(roleTitle)
-                        .font(.system(size: 11, weight: .semibold, design: .rounded))
-                        .foregroundStyle(ConsoleTheme.textMuted)
-
-                    if message.isStreaming {
-                        ProgressView()
-                            .controlSize(.mini)
-                            .tint(indicatorColor)
-                    }
-                }
-
-                Text(bodyText)
-                    .font(.system(size: 14, weight: .regular, design: .rounded))
-                    .foregroundStyle(ConsoleTheme.textPrimary)
-                    .textSelection(.enabled)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .padding(.horizontal, 12)
-            .padding(.vertical, 10)
-            .frame(maxWidth: 860, alignment: .leading)
-            .background(bubbleFill, in: RoundedRectangle(cornerRadius: 12, style: .continuous))
-            .overlay(
-                RoundedRectangle(cornerRadius: 12, style: .continuous)
-                    .strokeBorder(indicatorColor.opacity(0.36), lineWidth: 1)
-            )
-
-            if message.role != .user {
-                Spacer(minLength: 72)
-            }
-        }
-    }
-}
-
-private struct TimelineRow: View {
-    let entry: TimelineEntry
-
-    private var icon: String {
-        switch entry.level {
-        case .info:
-            return "info.circle"
-        case .action:
-            return "bolt.circle"
-        case .warning:
-            return "exclamationmark.triangle"
-        case .error:
-            return "xmark.octagon"
-        }
-    }
-
-    private var tint: Color {
-        switch entry.level {
-        case .info:
-            return ConsoleTheme.accent
-        case .action:
-            return ConsoleTheme.success
-        case .warning:
-            return ConsoleTheme.warning
-        case .error:
-            return ConsoleTheme.error
-        }
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
-            HStack(spacing: 7) {
-                Image(systemName: icon)
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundStyle(tint)
-
-                Text(entry.title)
-                    .font(.system(size: 12, weight: .semibold, design: .rounded))
-                    .foregroundStyle(ConsoleTheme.textPrimary)
-
-                Spacer(minLength: 6)
-
-                Text(entry.timestamp, style: .time)
-                    .font(.system(size: 10, weight: .regular, design: .monospaced))
-                    .foregroundStyle(ConsoleTheme.textMuted)
-            }
-
-            if let detail = entry.detail, !detail.isEmpty {
-                Text(detail)
-                    .font(.system(size: 11, weight: .regular, design: .rounded))
-                    .foregroundStyle(ConsoleTheme.textMuted)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-        .padding(.horizontal, 10)
-        .padding(.vertical, 8)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(ConsoleTheme.inlinePanel, in: RoundedRectangle(cornerRadius: 10, style: .continuous))
-        .overlay(
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .strokeBorder(ConsoleTheme.panelBorder, lineWidth: 1)
         )
     }
 }
